replace MemberOrganizaiton model in MemberModel and create a member.guard
