+++ libs/organization/src/lib/member/+state/index.ts
@@ -1,4 +1,4 @@
-//export * from './member.model';
+export * from './member.model';
 export * from './member.query';
 export * from './member.service';
 export * from './member.store';
