+++ libs/organization/src/lib/member/+state/member.model.ts
@@ -1,1 +1,17 @@
+export interface OrganizationMemberRequest {
+  email: string;
+  roles: string[];
+}
 
+export interface OrganizationMember extends OrganizationMemberRequest {
+  uid: string;
+  name?: string;
+  surname?: string;
+  avatar?: string;
+  role?: UserRole;
+}
+
+export const enum UserRole {
+  admin = 'admin',
+  member = 'member'
+}
